feat: fe - fabric이벤트 등록 및 소켓 연결 #83
- client to server 소켓 연결
